app/recipe_agent.py
import sys
import os
from typing import Optional, Type
from dotenv import load_dotenv
from pydantic import BaseModel
from pydantic_ai import Agent # Assuming this is the main class
# Add specific model/provider imports as originally intended, e.g.:
# from pydantic_ai.models.openai import OpenAIModel 
# from pydantic_ai.providers.openai import OpenAIProvider

from .models.recipe import Recipe # This is the critical import

load_dotenv()

class RecipeExtractorAgent:
    def __init__(self, output_model: Type[BaseModel] = Recipe):
        self.agent: Optional[Agent] = None # Or Optional[PydanticAI]
        self.output_model = output_model
        # self._initialize_agent() # Actual initialization logic would be here
        # print(f"RecipeExtractorAgent initialized with output model: {self.output_model.__name__}")

    def _initialize_agent(self):
        # Placeholder for agent initialization logic using pydantic_ai
        # This would configure the LLM provider, model, etc.
        # For example:
        # provider = OpenAIProvider(api_key=os.getenv("OPENAI_API_KEY"))
        # self.agent = Agent(model=OpenAIModel(model_name="gpt-3.5-turbo"), provider=provider, output_type=self.output_model)
        print("RecipeExtractorAgent: _initialize_agent called (placeholder)")

    async def extract_recipe_from_markdown(self, markdown_content: str) -> Optional[Recipe]:
        if not self.agent:
            print("PydanticAI agent is not initialized.")
            # self._initialize_agent() # Attempt to initialize if not already
            # if not self.agent: # Check again after attempt
            #     return None
            # For now, let's assume it should be pre-initialized or handle error
            return None # Or raise an exception
        
        # Placeholder: Actual call to pydantic_ai agent's run/arun method
        # result = await self.agent.arun(instruction="Extract recipe details.", data=markdown_content)
        # return result
        print(f"RecipeExtractorAgent: extract_recipe_from_markdown called (placeholder) for: {markdown_content[:50]}...")
        return Recipe(name="Placeholder Recipe", ingredients=["item1"], instructions=["step1"], image_url=None)

# print(f"--- [recipe_agent.py LOADED (reached end of file)] --- Name: {__name__} ---")