requirements.txt
fastapi
sqlalchemy
pydantic-ai
uvicorn[standard]
httpx
pytest
python-dotenv
crawl4ai
openai
google-generativeai
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
